Split interfaces and macros headers
Including the macros file also includes version.h which makes the
compilation dependent on this file. Removing this dependency makes
mc_rtc incremental builds much smoother.

This breaks client code for Observer and GlobalPlugin implementations
which are fairly recent.

For RobotModule, all known implementations only make use of
ROBOT_MODULE_API which is still in RobotModule.h

For controllers, we keep the macros in mc_controller.h and put the class
into MCController.h thus only requiring changes in mc_rtc
