+++ include/mc_control/fsm/Controller.h
@@ -4,8 +4,8 @@
 
 #pragma once
 
+#include <mc_control/MCController.h>
 #include <mc_control/fsm/Executor.h>
-#include <mc_control/mc_controller.h>
 #include <mc_tasks/EndEffectorTask.h>
 #include <mc_tasks/PostureTask.h>
