+++ src/mc_control/fsm/lib.cpp
@@ -4,4 +4,6 @@
 
 #include <mc_control/fsm/Controller.h>
 
+#include <mc_control/mc_controller.h>
+
 CONTROLLER_CONSTRUCTOR("FSM", mc_control::fsm::Controller)
