+++ src/mc_control/python/python_controller.h
@@ -2,7 +2,7 @@
  * Copyright 2015-2019 CNRS-UM LIRMM, CNRS-AIST JRL
  */
 
-#include <mc_control/mc_controller.h>
+#include <mc_control/MCController.h>
 
 namespace mc_rbdyn
 {
